Compile fixes

+++ Services/WCell.RealmServer/Entities/Character.Mechanics.cs
@@ -202,7 +202,7 @@ namespace WCell.RealmServer.Entities
 						// If the character current health is higher then the fall damage, the player survived the fall.
 						if (fallDamage < Health)
 						{
-							Achievements.CheckPossibleAchievementUpdates(AchievementCriteriaType.FallWithoutDying, m_fallStartHeight - m_position.Z * 100);
+							Achievements.CheckPossibleAchievementUpdates(AchievementCriteriaType.FallWithoutDying, (uint)(m_fallStartHeight - m_position.Z * 100));
 						}
 					//	DoEnvironmentalDamage(EnviromentalDamageType.Fall, fallDamage);
 					}
@@ -455,7 +455,7 @@ namespace WCell.RealmServer.Entities
 			base.DoEnvironmentalDamage(dmgType, amount);
 			if(!IsAlive)
 			{
-				Achievements.CheckPossibleAchievementUpdates(AchievementCriteriaType.DeathsFrom, dmgType, 1);
+				Achievements.CheckPossibleAchievementUpdates(AchievementCriteriaType.DeathsFrom, (uint)dmgType, 1);
 			}
 		}
 		#endregion
